pyproject.toml
[tool.poetry]
name = "serial-mailman"
version = "0.1.0"
description = ""
authors = ["Mikkel Roald-Arbøl <github.ggb9a@simplelogin.co>"]
readme = "README.md"
repository = "https://github.com/roaldarbol/mailman"
homepage = "https://github.com/roaldarbol/mailman"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
